+++ core/src/net/oschina/j2cache/ehcache/EhCacheProvider.java
@@ -16,9 +16,12 @@
 package net.oschina.j2cache.ehcache;
 
 import java.net.URL;
+import java.util.ArrayList;
+import java.util.Collection;
 import java.util.Properties;
 import java.util.concurrent.ConcurrentHashMap;
 
+import net.oschina.j2cache.CacheChannel;
 import net.oschina.j2cache.CacheObject;
 import net.sf.ehcache.config.CacheConfiguration;
 import org.slf4j.Logger;
@@ -51,6 +54,13 @@ public class EhCacheProvider implements CacheProvider {
 	@Override
 	public int level() {
 		return CacheObject.LEVEL_1;
+	}
+
+	@Override
+	public Collection<CacheChannel.Region> regions() {
+		Collection<CacheChannel.Region> regions = new ArrayList<>();
+		caches.forEach((k,c) -> regions.add(new CacheChannel.Region(k, c.size(), c.ttl())));
+		return regions;
 	}
 
     /**